requirements.txt
python-telegram-bot[job-queue]==21.1.1
Flask==3.0.3
requests==2.31.0
httpx==0.27.0
python-dotenv==1.0.1
